Adicionar componente CustomInputFormat para formatação de entrada de dados (telefone, CPF, RG e salário)
